Ya esta terminado el MVP ALV CHAVAAAAAAAL

+++ src/app/components/users/users.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ApipostService } from '../../services/apipost.service';
 import { UserModalComponent, UserUpdateComponent } from '../user-modal/user-modal.component';
-import { MatDialog } from '@angular/material';
+import { MatDialog, MatDialogConfig } from '@angular/material';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-users',
@@ -11,7 +12,9 @@ import { MatDialog } from '@angular/material';
 export class UsersComponent implements OnInit {
 
   userList: any = [];
-  constructor(private usersService:ApipostService, public dialog: MatDialog, private updateUser: UserUpdateComponent) { }
+  constructor(private usersService:ApipostService, 
+    public dialog: MatDialog
+    ) { }
 
   ngOnInit() {
     this.loadUsers();
@@ -21,20 +24,36 @@ export class UsersComponent implements OnInit {
     return this.usersService.getUsers().subscribe(
       (data: {}) =>{
         this.userList = data;
-        console.log(this.userList);
       }
     );
   }
 
   openModal() {
-    this.dialog.open(UserModalComponent);
-  }
+    const dialogRef1 = this.dialog.open(UserModalComponent);
 
-  deleteUser(user){
-    return this.usersService.deleteUsuario(user.id).subscribe((data: {}) => {
-      this.deleteEmployee(user.empleado.id);
-      console.log(data); 
+    dialogRef1.afterClosed().subscribe(dialogRef => {
+      if(dialogRef){
+        this.loadUsers();
+      }
     });
+
+  }
+//TODO correcto
+  deleteUser(user){
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Estas a punto de borrar un usuario',
+        message: '¿Estas seguro?'
+      }
+    });
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if(dialogResult){
+        return this.usersService.deleteUsuario(user.id).subscribe((data: {}) => {
+        this.deleteEmployee(user.empleado.id);
+        this.loadUsers(); 
+        });
+      }
+    })
   }
 
   deleteEmployee(id){
@@ -42,8 +61,29 @@ export class UsersComponent implements OnInit {
   }
 
   updateModal(user){
-    this.updateUser.getUser(user.id);
-    this.dialog.open(UserUpdateComponent);
+    
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+        data: {
+          title: 'Estas a punto de actualizar un usuario',
+          message: '¿Estas seguro?'
+        }
+      });
+
+      dialogRef.afterClosed().subscribe(dialogResult => {
+        if(dialogResult){
+          const dialogconfig = new MatDialogConfig();
+          dialogconfig.data = {
+            userId: user.id,
+            employeeId: user.empleado.id
+          };
+          const updatedialog = this.dialog.open(UserUpdateComponent, dialogconfig);
+          updatedialog.afterClosed().subscribe(dialogRef => {
+            if(dialogRef){
+              this.loadUsers();
+            }
+          });
+        }
+      });
   }
 
 }
